admin/requirements.txt
bert-serving-client==1.10.0
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
coverage==5.1
cryptography==2.9.2
elasticsearch==7.7.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.13.0
pyzmq==19.0.1
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-JSON==0.3.3
